Edited removeFromCartTest
