+++ src/main/resources/_testdata.properties
@@ -19,4 +19,3 @@ mobileNumber = 654321
 username = standard_user
 mobilePassword = secret_sauce
 productTitle = Sauce Labs Backpack
-productDetails = Sauce Labs Backpack carry.allTheThings() with the sleek, streamlined Sly Pack that melds uncompromising style with unequaled laptop and tablet protection. $29.99 REMOVE
