--- src/test/resources/ios-mobile-tests-suite.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="IOSMobileTests" parallel="classes" thread-count="6">
-    <test name="SauceLabsTests">
-        <classes>
-            <class name="com.solvd.laba.carina.mobiletests.SauceLabsTest">
-                <methods>
-                    <include name="testLoginUser"/>
-                    <include name="testAddToCart"/>
-                    <include name="testLogoutUser"/>
-                    <include name="testCheckoutUser"/>
-                    <include name="testAboutNavigation"/>
-                </methods>
-            </class>
-            <class name="com.solvd.laba.carina.mobiletests.IOSRemoveFromCartTest">
-                <methods>
-                    <include name="testRemoveFromCart"/>
-                </methods>
-            </class>
-        </classes>
-    </test>
-</suite>
