+++ src/test/resources/mobile-tests-suite.xml
@@ -1,20 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="AndroidMobileTests" parallel="classes" thread-count="6">
+<suite name="MobileTests" parallel="classes" thread-count="6">
     <test name="SauceLabsTests">
         <classes>
             <class name="com.solvd.laba.carina.mobiletests.SauceLabsTest">
                 <methods>
                     <include name="testLoginUser"/>
                     <include name="testAddToCart"/>
+                    <include name="testRemoveFromCart"/>
                     <include name="testLogoutUser"/>
                     <include name="testCheckoutUser"/>
                     <include name="testAboutNavigation"/>
-                </methods>
-            </class>
-            <class name="com.solvd.laba.carina.mobiletests.AndroidRemoveFromCartTest">
-                <methods>
-                    <include name="testRemoveFromCart"/>
                 </methods>
             </class>
         </classes>
